requirements.txt
# Core
django>=4.1.6

# Database
psycopg2-binary
dj-database-url

# Environment variables
python-dotenv

# Server
gunicorn
whitenoise

# LLM
ollama

# Development tools
django-extensions
